+++ src/main/java/mcpk/functions/sneaksprint/FunctionSneakSprint45.java
@@ -20,7 +20,7 @@ public class FunctionSneakSprint45 extends Function {
 			throws InvalidKeypressException {
 		Arguments args = new Arguments();
 		args.replace("duration", Math.abs(duration));
-		args.replace("facing", (float) (facing + 45));
+		args.replace("facing", facing + 45);
 		if (duration > 0) args.replace("forward", 1);
 		else if (duration < 0) args.replace("forward", -1);
