+++ src/main/java/mcpk/functions/sneaksprint/FunctionSneakSprintJump.java
@@ -20,7 +20,7 @@ public class FunctionSneakSprintJump extends Function {
 			throws DurationException {
 		Arguments args = new Arguments();
 		args.replace("duration", 1);
-		args.replace("facing", (float) facing);
+		args.replace("facing", facing);
 		if (duration > 0) args.replace("forward", 1);
 		else if (duration < 0) args.replace("forward", -1);
