+++ src/main/java/mcpk/functions/stop/FunctionStopAir.java
@@ -18,7 +18,7 @@ public class FunctionStopAir extends Function {
 	public void run(Player player, int duration, float facing, ArrayList<Character> modifiers, HashMap<String, Double> effects) throws DurationException, InvalidKeypressException {
 		Arguments args = new Arguments();
 		args.replace("duration", Math.abs(duration));
-		args.replace("facing", (float) facing);
+		args.replace("facing", facing);
 		
 		checkEffects(effects, args, duration);
 		checkNoModifiers(modifiers);
