+++ src/main/java/mcpk/functions/stop/FunctionStopJump.java
@@ -18,7 +18,7 @@ public class FunctionStopJump extends Function {
 	public void run(Player player, int duration, float facing, ArrayList<Character> modifiers, HashMap<String, Double> effects) throws DurationException, InvalidKeypressException {
 		Arguments args = new Arguments();
 		args.replace("duration", 1);
-		args.replace("facing", (float) facing);
+		args.replace("facing", facing);
 		
 		checkEffects(effects, args, duration);
 		checkNoModifiers(modifiers);
