+++ src/main/java/mcpk/functions/walk/FunctionWalkJump.java
@@ -18,7 +18,7 @@ public class FunctionWalkJump extends Function {
 	public void run(Player player, int duration, float facing, ArrayList<Character> modifiers, HashMap<String, Double> effects) throws DurationException {
 		Arguments args = new Arguments();
 		args.replace("duration", 1);
-		args.replace("facing", (float) (facing));
+		args.replace("facing", facing);
 		if (duration > 0) args.replace("forward", 1);
 		else if (duration < 0) args.replace("forward", -1);
