+++ src/main/java/mcpk/functions/walk/FunctionWalkJump45.java
@@ -18,7 +18,7 @@ public class FunctionWalkJump45 extends Function {
 	public void run(Player player, int duration, float facing, ArrayList<Character> modifiers, HashMap<String, Double> effects) throws InvalidKeypressException {
 		Arguments args = new Arguments();
 		args.replace("duration", 1);
-		args.replace("facing", (float) (facing + 45));
+		args.replace("facing", facing + 45);
 		if (duration > 0) args.replace("forward", 1);
 		else if (duration < 0) args.replace("forward", -1);
